Moving RetryableOperation to separate class.
